add confirm to deletion

+++ src/stock/StockItem.tsx
@@ -16,24 +16,26 @@ export const StockItem = (props) => {
       : props.info.description;
 
   let handleDelete = async (id) => {
-    try {
-      let res = await fetch("http://localhost:3001/remove", {
-        method: "DELETE",
-        headers: {
-          Accept: "application/json",
-          "Content-type": "application/json",
-        },
-        body: JSON.stringify({
-          id: id,
-        }),
-      });
-      let resJson = await res.json();
-      if (res.status === 200) {
-        setIsActive(false);
-      } else {
+    if (confirm(`Are you sure you want to delete Stock ${id}`)) {
+      try {
+        let res = await fetch("http://localhost:3001/remove", {
+          method: "DELETE",
+          headers: {
+            Accept: "application/json",
+            "Content-type": "application/json",
+          },
+          body: JSON.stringify({
+            id: id,
+          }),
+        });
+        let resJson = await res.json();
+        if (res.status === 200) {
+          setIsActive(false);
+        } else {
+        }
+      } catch (err) {
+        console.log(err);
       }
-    } catch (err) {
-      console.log(err);
     }
   };
 
@@ -51,7 +53,7 @@ export const StockItem = (props) => {
               </Button>
             </Link>
             <Button onClick={() => handleDelete(props.info.stock_id)}>
-              <FontAwesomeIcon icon={faTrash}/>
+              <FontAwesomeIcon icon={faTrash} />
             </Button>
           </td>
         </tr>
